Issue #13, moved all actions into actions subfolder, added actions/upstreamKeyActions git commit -am Issue
