+++ .gitignore
@@ -5,3 +5,4 @@ package-lock.json
 dist/
 DEBUG-*
 /.yarn
+/docs
