+++ README.md
@@ -3,7 +3,6 @@
 [![License](https://img.shields.io/github/license/bitfocus/companion-module-osee-gostream)](https://github.com/bitfocus/companion-module-osee-gostream/blob/main/LICENSE)
 [![Version](https://img.shields.io/github/v/release/bitfocus/companion-module-osee-gostream)](https://github.com/bitfocus/companion-module-osee-gostream/releases)
 
-
 This module is created for use with GoStream Devices.
 
 - [Read the docs](./How%20to%20connect%20to%20computer%20software%20and%20companion.pdf)
