+++ package.json
@@ -11,7 +11,8 @@
 		"build": "rimraf dist && yarn build::main",
 		"build::main": "tsc -p tsconfig.build.json",
 		"build::dev": "tsc -p tsconfig.build.json --watch",
-		"prepare": "husky"
+		"prepare": "husky",
+		"doc": "npx typedoc --out docs src"
 	},
 	"repository": {
 		"type": "git",
@@ -22,12 +23,14 @@
 	},
 	"dependencies": {
 		"@companion-module/base": "~1.11.0",
-		"socket.io-client": "^4.5.1"
+		"socket.io-client": "^4.5.1",
+		"typescript-eslint": "^8.8.1"
 	},
 	"devDependencies": {
 		"@companion-module/tools": "^2.1.0",
-		"prettier": "^3.3.3",
 		"husky": "^9.1.6",
+		"prettier": "^3.3.3",
+		"typedoc": "^0.26.8",
 		"typescript": "~5.5.4"
 	}
 }
